chart用の体重データ

+++ pppppp/Model/ChartsWeightItem.swift
@@ -2,6 +2,6 @@ import Foundation
 
 struct ChartsWeightItem: Identifiable {
     var date: String
-    var weight: Int
+    var weight: Float
     var id = UUID()
 }
